+++ backend/src/main/java/com/wms/domain/inputWarehouse/service/impl/InputWarehouseServiceImpl.java
@@ -39,10 +39,6 @@ public class InputWarehouseServiceImpl implements InputWarehouseService {
             purchaseSheetRepository;
     private final InputWarehouseDetailService inputWarehouseDetailService;
     private final UserFindService userFindService;
-    private final InputWarehouseDetailRepository inputWarehouseDetailRepository;
-    private final InventoryService inventoryService;
-    private final PurchaseDetailService purchaseDetailService;
-    private final PurchaseSheetService purchaseSheetService;
 
     @Override
     @Transactional
@@ -70,34 +66,4 @@ public class InputWarehouseServiceImpl implements InputWarehouseService {
                 .collect(Collectors.toList());
     }
 
-    @Override
-    @Transactional
-    public void completeInputWarehouse(List<Long> inputWarehouseDetailIds) {
-        // 1. 해당 입고 상세 정보 조회
-        List<InputWarehouseDetail> inputWarehouseDetails = inputWarehouseDetailRepository.findAllById(inputWarehouseDetailIds);
-
-        if (inputWarehouseDetails.isEmpty()) {
-            throw new InputWarehouseDetailException(InputWarehouseDetailExceptionResponseCode.INPUT_WAREHOUSE_DETAIL_NOT_FOUND, "입고 상세 정보를 찾을 수 없습니다.");
-        }
-
-        // 2. 재고 업데이트
-        for (InputWarehouseDetail detail : inputWarehouseDetails) {
-            inventoryService.updateInventory(detail);
-
-            // 발주 상세 상태 변경
-            purchaseDetailService.completePurchaseDetail(detail.getPurchaseDetail().getId());
-        }
-
-        // 3. 발주서 상태 업데이트
-        PurchaseSheet purchaseSheet = inputWarehouseDetails.get(0).getInputWarehouse()
-                .getPurchaseSheet();
-
-        if (purchaseSheet == null) {
-            throw new InputWarehouseException(
-                    InputWarehouseExceptionResponseCode.INPUT_WAREHOUSE_NOT_FOUND, "입고 상세 정보에 해당하는 발주서를 찾을 수 없습니다."
-            );
-        }
-
-        purchaseSheetService.updatePurchaseSheetStatus(purchaseSheet);
-    }
 }
