+++ backend/src/main/java/com/wms/domain/inputWarehouseDetail/service/impl/InputWarehouseDetailServiceImpl.java
@@ -11,12 +11,16 @@ import com.wms.domain.inputWarehouseDetail.dto.response.InputWarehouseDetailsRes
 import com.wms.domain.inputWarehouseDetail.entity.InputWarehouseDetail;
 import com.wms.domain.inputWarehouseDetail.repository.InputWarehouseDetailRepository;
 import com.wms.domain.inputWarehouseDetail.service.InputWarehouseDetailService;
+import com.wms.domain.inventory.service.InventoryService;
 import com.wms.domain.item.entity.Item;
 import com.wms.domain.item.repository.ItemRepository;
 import com.wms.domain.lot.entity.Lot;
 import com.wms.domain.lot.repository.LotRepository;
 import com.wms.domain.purchaseDetail.entity.PurchaseDetail;
 import com.wms.domain.purchaseDetail.repository.PurchaseDetailRepository;
+import com.wms.domain.purchaseDetail.service.PurchaseDetailService;
+import com.wms.domain.purchaseSheet.entity.PurchaseSheet;
+import com.wms.domain.purchaseSheet.service.PurchaseSheetService;
 import com.wms.domain.rack.entity.Rack;
 import com.wms.domain.rack.repository.RackRepository;
 import com.wms.domain.warehouse.entity.Warehouse;
@@ -44,6 +48,9 @@ public class InputWarehouseDetailServiceImpl implements InputWarehouseDetailServ
     private final AreaRepository areaRepository;
     private final RackRepository rackRepository;
     private final CellRepository cellRepository;
+    private final InventoryService inventoryService;
+    private final PurchaseDetailService purchaseDetailService;
+    private final PurchaseSheetService purchaseSheetService;
 
     @Override
     @Transactional
@@ -99,5 +106,36 @@ public class InputWarehouseDetailServiceImpl implements InputWarehouseDetailServ
         return inputWarehouseDetails.stream()
                 .map(InputWarehouseDetailsResponseDTO2::fromEntity)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    @Transactional
+    public void completeInputWarehouse(List<Long> inputWarehouseDetailIds) {
+        // 1. 해당 입고 상세 정보 조회
+        List<InputWarehouseDetail> inputWarehouseDetails = inputWarehouseDetailRepository.findAllById(inputWarehouseDetailIds);
+
+        if (inputWarehouseDetails.isEmpty()) {
+            throw new InputWarehouseDetailException(InputWarehouseDetailExceptionResponseCode.INPUT_WAREHOUSE_DETAIL_NOT_FOUND, "입고 상세 정보를 찾을 수 없습니다.");
+        }
+
+        // 2. 재고 업데이트
+        for (InputWarehouseDetail detail : inputWarehouseDetails) {
+            inventoryService.updateInventory(detail);
+
+            // 발주 상세 상태 변경
+            purchaseDetailService.completePurchaseDetail(detail.getPurchaseDetail().getId());
+        }
+
+        // 3. 발주서 상태 업데이트
+        PurchaseSheet purchaseSheet = inputWarehouseDetails.get(0).getInputWarehouse()
+                .getPurchaseSheet();
+
+        if (purchaseSheet == null) {
+            throw new InputWarehouseException(
+                    InputWarehouseExceptionResponseCode.INPUT_WAREHOUSE_NOT_FOUND, "입고 상세 정보에 해당하는 발주서를 찾을 수 없습니다."
+            );
+        }
+
+        purchaseSheetService.updatePurchaseSheetStatus(purchaseSheet);
     }
 }
