+++ backend/src/main/java/com/wms/domain/sell/controller/SellController.java
@@ -1,4 +1,14 @@
 package com.wms.domain.sell.controller;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api/v1/sell")
 public class SellController {
+
 }
